// api/playauto_api_type.ts
export const shopDataNameInfo: { [key: string]: string } = {
    "A001": "옥션",
    "A522": "옥션2.0",
    "A006": "G마켓",
    "A523": "G마켓2.0",
    "A077": "스마트스토어",
    "A027": "인터파크 - 오픈마켓",
    "A112": "11번가",
    "A113": "11번가(단일상품)",
    "B955": "마이소",
    "B118": "메이크샵",
    "B378": "쿠팡",
    "B956": "티몬",
    "B430": "위메프",
    "B719": "위메프2.0",
    "C003": "Qoo10(sg)",
    "C010": "Qoo10(jp)",
    "A088": "멸치쇼핑",
    "B579": "NH마켓",
    "A011": "GS SHOP",
    "A012": "롯데닷컴",
    "A004": "CJmall",
    "B008": "half club",
    "A131": "현대H몰 (Hmall)",
    "B009": "LOTTE DEPARTMENT STORE",
    "B617": "오가게",
    "A017": "Akmall",
    "A065": "iSTYLE 24",
    "A024": "WIZWID",
    "B394": "the galleria",
    "A118": "ns홈쇼핑",
    "B666": "MUSINSA",
    "B733": "PLAYER",
    "B612": "LOTTE MART",
    "B614": "홈앤쇼핑",
    "B007": "패션플러스",
    "B957": "W컨셉",
    "A032": "SSG(통합) - 신세계몰, SSG(이마트몰), 이마트트레이더스몰",
    "B609": "다이소몰",
    "B048": "1300k",
    "B041": "텐바이텐",
    "B051": "바보사랑",
    "B599": "인터파크 도서",
    "B046": "예스24",
    "B888": "후추통",
    "B572": "골핑(B2C)",
    "B087": "sbs골프닷컴",
    "A132": "현대H몰(백화점)",
    "B325": "이지웰",
    "A085": "SSG(신세계백화점)",
    "B701": "삼성카드쇼핑",
    "B690": "K쇼핑",
    "B771": "제로투세븐",
    "A100": "롯데아이몰",
    "B661": "이랜드몰",
    "B693": "도매꾹",
    "B668": "도매창고",
    "B730": "오너클랜",
    "B669": "머스트잇",
    "B682": "브리치",
    "B685": "공영홈쇼핑",
    "B688": "카카오톡스토어",
    "B953": "POOM",
    "B694": "핫트랙스",
    "A133": "더현대닷컴",
    "B696": "하이마트",
    "B502": "홈플러스",
    "B959": "오늘의집",
    "A524": "롯데온",
    "B880": "펀앤쇼핑",
}
// ...
export interface IPAApi {
    amp_key: String,
    /**
     *EMP 기준 사용자별 고유 코드
     *
     * @author Kuhave
     * @memberof IPAApi
     */
    amp_program_code: string;
    /**
     *사용자별 고유 타입
     *
     * @author Kuhave
     * @memberof IPAApi
     */
    amp_program_type: string;
    /**
     *사용자별 기본 디렉토리 경로(필수X)
     *
     * @author Kuhave
     * @memberof IPAApi
     */
    amp_program_homedir: string;
    /**
     *사용자별 기본 URL 경로(필수X)
     *
     * @author Kuhave
     * @memberof IPAApi
     */
    amp_program_homeurl: string;
}
// ...
export interface IPAConfig {
    ProdUseDenyKeyword: boolean | string;
    ProdUseDenyKeywordString: string;
    ProdUseCancelMsg: boolean | string;
    ProdUseCancelMsgString: string;
    ProdSearchDate: string;
    ProdUseEditImg: boolean | string;
    OrderSearchDate: string; //주문 및 문의 수집 기간
}
export interface IPADataDataSet {
    api: IPAApi[];
    config: IPAConfig[]
}

export interface IPADataDataSet10 extends IPADataDataSet {
    data_opt: IPADataOpt[];
    data_slave: IPADataSlave[];
    data_set: [],
    data: IPAData[],
}
export interface IPADataDataSet3 extends IPADataDataSet {
    confirm_order: IPAConfirmOrder[];
}

export interface IPAConfirmOrder {
    confirm_doit: boolean
}

export interface IPADShopInfo<T extends IPADataDataSet> {
    action: number[];
    account_code: string;
    site_code: string;
    dll_code: string;
    site_name: string;
    encoding: string;
    id: string;
    id2: string;
    pw: string;
    etc: string;
    etc2: string;
    etc3: string;
    etc4: string;
    etc5: string;
    etc6: string;
    global_yn: string;
    prod_codes: string[];
    DataDataSet: T;
    DataDataSetFileName: string;
    syncList: string[] | string;
    fromState: string;
    SetNo: string;
    Desc: string;
    dummy1: number;
}